renamed StacAssetExplorer to StacExplorer
