+++ sites/geohub/src/components/stac/StacExplorer.svelte
